Renamed icinga.php to nagios.php, changed references in code
