+++ iplock.php
@@ -36,7 +36,7 @@ function validate_ip_against_config($iplist) {
 // @codingStandardsIgnoreEnd
 
     // Require library for icinga responses
-    require_once('icinga.php');
+    require_once('nagios.php');
     // Validate remote IP against safe list.
     if (remoteip_in_list($iplist)) {
         return;
